chore(server): reorg functions & CLI wip
wip


wip


wip
